Changes for database apis to retrieve warning
