--- imports/api/sarifFiles.js
@@ -1,7 +1,0 @@
-
-import { Mongo } from "meteor/mongo"; 
-
-export const SarifFiles = new Mongo.Collection("SarifFiles");
-
-
-
