+++ imports/ui/prototype1/about.js
@@ -1,6 +1,6 @@
 // Libraries
 
-import {Projects} from "../../api/projects.js";
+import {Projects} from "../../../collections/projects.js";
 import {Template} from "meteor/templating";
 
 import "../table/table.html";
