+++ imports/ui/prototype2/prototype2.js
@@ -1,6 +1,6 @@
 // Libraries
 
-import {Projects} from "../../api/projects.js";
+import {Projects} from "../../../collections/projects.js";
 import {Template} from "meteor/templating";
 
 // Prototype 2
