+++ server/main.js
@@ -1,2 +1,2 @@
-import "../imports/api/sarifFiles.js";
-import "../imports/api/projects.js";
+import {Projects} from "../collections/projects.js";
+import {SarifFiles} from "../collections/sarifFiles.js";
